requirements.txt
requests==2.19.1
python-dotenv==0.9.1
tornado==5.1
apscheduler
tushare
bs4